move counterparty logic and msgs into packetserver
