+++ modules/core/02-client/keeper/keeper.go
@@ -304,25 +304,6 @@ func (k *Keeper) GetLatestClientConsensusState(ctx sdk.Context, clientID string)
 	return k.GetClientConsensusState(ctx, clientID, clientModule.LatestHeight(ctx, clientID))
 }
 
-// SetCounterparty sets the Counterparty for a given client identifier.
-func (k *Keeper) SetCounterparty(ctx sdk.Context, clientID string, counterparty types.Counterparty) {
-	bz := k.cdc.MustMarshal(&counterparty)
-	k.ClientStore(ctx, clientID).Set([]byte(types.CounterpartyKey), bz)
-}
-
-// GetCounterparty gets the Counterparty for a given client identifier.
-func (k *Keeper) GetCounterparty(ctx sdk.Context, clientID string) (types.Counterparty, bool) {
-	store := k.ClientStore(ctx, clientID)
-	bz := store.Get([]byte(types.CounterpartyKey))
-	if len(bz) == 0 {
-		return types.Counterparty{}, false
-	}
-
-	var counterparty types.Counterparty
-	k.cdc.MustUnmarshal(bz, &counterparty)
-	return counterparty, true
-}
-
 // GetCreator returns the creator of the client.
 func (k *Keeper) GetCreator(ctx sdk.Context, clientID string) (string, bool) {
 	bz := k.ClientStore(ctx, clientID).Get([]byte(types.CreatorKey))
