+++ modules/core/02-client/types/client.go
@@ -12,7 +12,6 @@ import (
 
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 
-	commitmenttypes "github.com/cosmos/ibc-go/v9/modules/core/23-commitment/types/v2"
 	host "github.com/cosmos/ibc-go/v9/modules/core/24-host"
 	"github.com/cosmos/ibc-go/v9/modules/core/exported"
 )
@@ -63,27 +62,6 @@ func (ics IdentifiedClientStates) Swap(i, j int) { ics[i], ics[j] = ics[j], ics[
 func (ics IdentifiedClientStates) Sort() IdentifiedClientStates {
 	sort.Sort(ics)
 	return ics
-}
-
-// NewCounterparty creates a new Counterparty instance
-func NewCounterparty(clientID string, merklePathPrefix *commitmenttypes.MerklePath) Counterparty {
-	return Counterparty{
-		ClientId:         clientID,
-		MerklePathPrefix: merklePathPrefix,
-	}
-}
-
-// Validate validates the Counterparty
-func (c Counterparty) Validate() error {
-	if err := host.ClientIdentifierValidator(c.ClientId); err != nil {
-		return err
-	}
-
-	if c.MerklePathPrefix.Empty() {
-		return errorsmod.Wrap(ErrInvalidCounterparty, "prefix cannot be empty")
-	}
-
-	return nil
 }
 
 // NewConsensusStateWithHeight creates a new ConsensusStateWithHeight instance
