+++ modules/core/24-host/channel_keys.go
@@ -2,6 +2,8 @@ package host
 
 import "fmt"
 
+var KeyChannelStorePrefix = []byte("channels")
+
 const (
 	KeyChannelEndPrefix        = "channelEnds"
 	KeyChannelPrefix           = "channels"
