+++ modules/core/packet-server/types/errors.go
@@ -1,0 +1,10 @@
+package types
+
+import (
+	errorsmod "cosmossdk.io/errors"
+)
+
+// IBC packet server sentinel errors
+var (
+	ErrInvalidCounterparty = errorsmod.Register(SubModuleName, 1, "invalid counterparty")
+)
